Misc fixes to admin rooms and sending of SMS
This was broken as part of the migration, and there were flaws in the admin room processing. This now tries to figure out if the room is a 1:1 with a phone number or an admin room. 1:1 chats with phones require the bridge user to avoid load on the bridge side. The ghost will now try to invite the bridge (which auto-joins) if it can't see the bridge bot in the room.